pyproject.toml
[project]
name = "ai-cli-clone"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "openai-agents>=0.3.3",
    "python-dotenv>=1.1.1",
    "requests>=2.32.5",
    "rich>=14.2.0",
]
